+++ examples/test.sh
@@ -1,0 +1,13 @@
+#!/bin/sh
+
+echo "hi"
+
+3
+4
+56
+
+hhhhh
+printf "hello world\n"
+echo "bye"
+sum=$((1 + 2))
+echo "sum: $sum"